Planning React Tutorial
Going well so far. Trying to make React approachable using the most widely available build tool: Vite. A couple adjustments to the main Planning.md file, and I have yet to write the ten short-answer questions for the final exam. But I'm hoping this will all be done by the end of February.