Session Push - 02/06/2024 TUE 0907
I forgot to push the code yesterday, so this encompasses both today's and yesterday's revisions. I discovered a couple of new developments, namely that the useEffect hook is NOT required to make an API call, and Axios is probably a better way of calling also. I want the students to start off with the ability to call a random verse, and then give them the tools they need to be able to call the API for a specific verse without explicitly telling them how to do it. Still tweaking with it.